Move helper func to experimental namespace.
