+++ clang/test/dpct/texture/texture_object_bindless_image.cu
@@ -23,7 +23,7 @@ int main() {
   size_t w, h, sizeInBytes, w_offest_src, h_offest_src, w_offest_dest, h_offest_dest;
   unsigned int flag;
   cudaExtent e;
-  // CHECK: dpct::image_mem_p pArr, pArr_src;
+  // CHECK: dpct::experimental::image_mem_ptr pArr, pArr_src;
   cudaArray_t pArr, pArr_src;
   // TODO: need support.
   // cudaMipmappedArray_t pMipMapArr;
@@ -33,62 +33,71 @@ int main() {
   cudaMalloc3DArray(&pArr, &desc, e);
   // CHECK: pArr = new sycl::ext::oneapi::experimental::image_mem(sycl::ext::oneapi::experimental::image_descriptor({w, h}, desc.get_channel_order(), desc.get_channel_type()), q_ct1);
   cudaMallocArray(&pArr, &desc, w, h);
-  // CHECK: desc = dpct::get_channel(pArr);
+  // CHECK: desc = dpct::experimental::get_channel(pArr);
   // CHECK-NEXT: e = pArr->get_range();
   // CHECK-NEXT: flag = 0;
   cudaArrayGetInfo(&desc, &e, &flag, pArr);
-  // CHECK: dpct::dpct_memcpy(pArr_src, w_offest_src, h_offest_src, pArr, w_offest_dest, h_offest_dest, w, h, q_ct1);
+  // CHECK: dpct::experimental::dpct_memcpy(pArr_src, w_offest_src, h_offest_src, pArr, w_offest_dest, h_offest_dest, w, h, q_ct1);
   cudaMemcpy2DArrayToArray(pArr, w_offest_dest, h_offest_dest, pArr_src,
                            w_offest_src, h_offest_src, w, h,
                            cudaMemcpyHostToDevice);
-  // CHECK: dpct::dpct_memcpy(pArr, w_offest_src, h_offest_src, input, w, w, h, q_ct1);
+  // CHECK: dpct::experimental::dpct_memcpy(pArr, w_offest_src, h_offest_src, input, w, w, h, q_ct1);
   cudaMemcpy2DFromArray(input, w, pArr, w_offest_src, h_offest_src, w, h,
                         cudaMemcpyHostToDevice);
-  // CHECK: dpct::async_dpct_memcpy(pArr, w_offest_src, h_offest_src, input, w, w, h, q_ct1);
+  // CHECK: dpct::experimental::async_dpct_memcpy(pArr, w_offest_src, h_offest_src, input, w, w, h, q_ct1);
   cudaMemcpy2DFromArrayAsync(input, w, pArr, w_offest_src, h_offest_src, w, h,
                              cudaMemcpyHostToDevice);
-  // CHECK: dpct::dpct_memcpy(input, pArr, w_offest_dest, h_offest_dest, w, w, h, q_ct1);
+  // CHECK: dpct::experimental::dpct_memcpy(input, pArr, w_offest_dest, h_offest_dest, w, w, h, q_ct1);
   cudaMemcpy2DToArray(pArr, w_offest_dest, h_offest_dest, input, w, w, h,
                       cudaMemcpyHostToDevice);
-  // CHECK: dpct::async_dpct_memcpy(input, pArr, w_offest_dest, h_offest_dest, w, w, h, q_ct1);
+  // CHECK: dpct::experimental::async_dpct_memcpy(input, pArr, w_offest_dest, h_offest_dest, w, w, h, q_ct1);
   cudaMemcpy2DToArrayAsync(pArr, w_offest_dest, h_offest_dest, input, w, w, h,
                            cudaMemcpyHostToDevice);
-  // CHECK: dpct::dpct_memcpy(pArr_src, w_offest_src, h_offest_src, pArr, w_offest_dest, h_offest_dest, w * h, q_ct1);
+  // CHECK: dpct::experimental::dpct_memcpy(pArr_src, w_offest_src, h_offest_src, pArr, w_offest_dest, h_offest_dest, w * h, q_ct1);
   cudaMemcpyArrayToArray(pArr, w_offest_dest, h_offest_dest, pArr_src,
                          w_offest_src, h_offest_src, w * h,
                          cudaMemcpyHostToDevice);
-  // CHECK: dpct::dpct_memcpy(pArr, w_offest_src, h_offest_src, input, w * h, q_ct1);
+  // CHECK: dpct::experimental::dpct_memcpy(pArr, w_offest_src, h_offest_src, input, w * h, q_ct1);
   cudaMemcpyFromArray(input, pArr, w_offest_src, h_offest_src, w * h,
                       cudaMemcpyHostToDevice);
-  // CHECK: dpct::async_dpct_memcpy(pArr, w_offest_src, h_offest_src, input, w * h, q_ct1);
+  // CHECK: dpct::experimental::async_dpct_memcpy(pArr, w_offest_src, h_offest_src, input, w * h, q_ct1);
   cudaMemcpyFromArrayAsync(input, pArr, w_offest_src, h_offest_src, w * h,
                            cudaMemcpyHostToDevice);
-  // CHECK: dpct::dpct_memcpy(input, pArr, w_offest_dest, h_offest_dest, w * h, q_ct1);
+  // CHECK: dpct::experimental::dpct_memcpy(input, pArr, w_offest_dest, h_offest_dest, w * h, q_ct1);
   cudaMemcpyToArray(pArr, w_offest_dest, h_offest_dest, input, w * h,
                     cudaMemcpyHostToDevice);
-  // CHECK: dpct::async_dpct_memcpy(input, pArr, w_offest_dest, h_offest_dest, w * h, q_ct1);
+  // CHECK: dpct::experimental::async_dpct_memcpy(input, pArr, w_offest_dest, h_offest_dest, w * h, q_ct1);
   cudaMemcpyToArrayAsync(pArr, w_offest_dest, h_offest_dest, input, w * h,
                          cudaMemcpyHostToDevice);
 
-  // CHECK: dpct::image_data resDesc1, resDesc2, resDesc3, resDesc4;
-  cudaResourceDesc resDesc1, resDesc2, resDesc3, resDesc4;
-  // CHECK: resDesc1.set_data(pArr);
-  resDesc1.resType = cudaResourceTypeArray;
+  // CHECK: dpct::image_data resDesc0, resDesc1, resDesc2, resDesc3, resDesc4;
+  cudaResourceDesc resDesc0, resDesc1, resDesc2, resDesc3, resDesc4;
+  // CHECK: resDesc0.set_data_type(dpct::image_data_type::matrix);
+  resDesc0.resType = cudaResourceTypeArray;
+  // CHECK: resDesc1.set_data_ptr(pArr);
   resDesc1.res.array.array = pArr;
   // TODO: need support.
-  // resDesc2.resType = cudaResourceTypeMipmappedArray;
+  // resDesc0.resType = cudaResourceTypeMipmappedArray;
   // resDesc2.res.mipmap.mipmap = pMipMapArr;
-  // CHECK: resDesc3.set_data(input, sizeInBytes, desc);
-  resDesc3.resType = cudaResourceTypeLinear;
+  // CHECK: resDesc0.set_data_type(dpct::image_data_type::linear);
+  resDesc0.resType = cudaResourceTypeLinear;
+  // CHECK: resDesc3.set_data_ptr(input);
   resDesc3.res.linear.devPtr = input;
+  // CHECK: resDesc3.set_channel(desc);
   resDesc3.res.linear.desc = desc;
+  // CHECK: resDesc3.set_x(sizeInBytes);
   resDesc3.res.linear.sizeInBytes = sizeInBytes;
-  // CHECK: resDesc4.set_data(input, w, h, sizeInBytes, desc);
-  resDesc4.resType = cudaResourceTypePitch2D;
+  // CHECK: resDesc0.set_data_type(dpct::image_data_type::pitch);
+  resDesc0.resType = cudaResourceTypePitch2D;
+  // CHECK: resDesc4.set_data_ptr(input);
   resDesc4.res.pitch2D.devPtr = input;
+  // CHECK: resDesc4.set_channel(desc);
   resDesc4.res.pitch2D.desc = desc;
+  // CHECK: resDesc4.set_x(w);
   resDesc4.res.pitch2D.width = w;
+  // CHECK: resDesc4.set_y(h);
   resDesc4.res.pitch2D.height = h;
+  // CHECK: resDesc4.set_pitch(sizeInBytes);
   resDesc4.res.pitch2D.pitchInBytes = sizeInBytes;
 
   // CHECK: dpct::sampling_info texDesc1, texDesc2, texDesc3, texDesc4;
@@ -112,13 +121,13 @@ int main() {
 
   // CHECK: sycl::ext::oneapi::experimental::sampled_image_handle tex;
   cudaTextureObject_t tex;
-  // CHECK: tex = dpct::create_bindless_image(resDesc1, texDesc1);
+  // CHECK: tex = dpct::experimental::create_bindless_image(resDesc1, texDesc1);
   cudaCreateTextureObject(&tex, &resDesc1, &texDesc1, NULL);
-  // CHECK: desc = dpct::get_channel(pArr);
+  // CHECK: desc = dpct::experimental::get_channel(pArr);
   cudaGetChannelDesc(&desc, pArr);
-  // CHECK: resDesc1 = dpct::get_data(tex);
+  // CHECK: resDesc1 = dpct::experimental::get_data(tex);
   cudaGetTextureObjectResourceDesc(&resDesc1, tex);
-  // CHECK: texDesc1 = dpct::get_sampling_info(tex);
+  // CHECK: texDesc1 = dpct::experimental::get_sampling_info(tex);
   cudaGetTextureObjectTextureDesc(&texDesc1, tex);
   // CHECK: q_ct1.parallel_for(
   // CHECK-NEXT: sycl::nd_range<3>(sycl::range<3>(1, 1, 1), sycl::range<3>(1, 1, 1)),
@@ -126,7 +135,7 @@ int main() {
   // CHECK-NEXT:   kernel(tex);
   // CHECK-NEXT: });
   kernel<<<1, 1>>>(tex);
-  // CHECK: dpct::destroy_bindless_image(tex, q_ct1);
+  // CHECK: dpct::experimental::destroy_bindless_image(tex, q_ct1);
   cudaDestroyTextureObject(tex);
   // CHECK: sycl::ext::oneapi::experimental::free_image_mem(pArr->get_handle(), dpct::get_in_order_queue());
   cudaFreeArray(pArr);
